+++ ninja-core/src/main/java/ninja/utils/NinjaPropertiesImpl.java
@@ -26,6 +26,7 @@ import org.apache.commons.configuration.Configuration;
 import org.apache.commons.configuration.ConfigurationConverter;
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.commons.configuration.PropertiesConfiguration;
+import org.apache.commons.configuration.reloading.FileChangedReloadingStrategy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -77,7 +78,7 @@ public class NinjaPropertiesImpl implements NinjaProperties {
         Configuration prefixedDefaultConfiguration = null;
 
         // (Optional) Config set via a system property
-        Configuration externalConfiguration = null;
+        PropertiesConfiguration externalConfiguration = null;
 
         // (Optional) Config of prefixed mode corresponding to current mode (eg.
         // %test.myproperty=...)
@@ -135,6 +136,11 @@ public class NinjaPropertiesImpl implements NinjaProperties {
                 throw new RuntimeException(errorMessage);
 
             } else {
+
+                // allow the external configuration to be reloaded at
+                // runtime based on detected file changes
+                externalConfiguration
+                        .setReloadingStrategy(new FileChangedReloadingStrategy());
 
                 // Copy special prefix of mode to parent configuration
                 // By convention it will be something like %test.myproperty