+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,3 +1,9 @@
+Version 3.x.x
+=============
+
+* 2014-10-01 Add automatic hot reloading support for external (`-Dninja.external.conf`) configuration files for all runtime modes (gitblit)
+* 2014-10-01 Add automatic hot reloading support for language `messages` files in **dev** mode (gitblit)
+
 Version 3.3.3
 =============